requirements.txt
# Stable Diffusion Image Generator Dependencies
# Core dependencies for AI image generation application

# Deep Learning Framework
# Note: For CUDA support, install with: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Hugging Face Ecosystem - Updated for SD 3.5 support
diffusers>=0.31.0
transformers>=4.44.0
accelerate>=0.34.0

# Image Processing
Pillow>=9.0.0
opencv-python>=4.7.0

# Scientific Computing
numpy>=1.21.0
scipy>=1.9.0

# Utility Libraries
tqdm>=4.64.0
requests>=2.28.0
safetensors>=0.3.0

# System Monitoring
psutil>=5.9.0

# GUI Framework (PyQt5 Only)
PyQt5>=5.15.0
PyQt5-Qt5>=5.15.0

# 修复Windows/Linux下剪贴板等平台相关问题的插件
pyqt5-plugins>=5.15.0.0

# 军事目标检测和标注依赖
ultralytics>=8.0.0
opencv-python>=4.7.0
PyYAML>=6.0

# 数据处理和分析
pandas>=1.5.0
matplotlib>=3.6.0
seaborn>=0.12.0

# 可选：GPU加速支持
# 如需GPU支持，请手动安装：
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
